optimization and adding documents
