+++ .env.default
@@ -1,0 +1,19 @@
+# Environment Variables
+
+# Database Configuration
+DB_HOST=localhost
+DB_PORT=27017
+DB_NAME=userdb
+
+# Application Configuration
+APP_PORT=3000
+APP_ENV=development
+USER_SERVICE_URL=http://localhost:8081
+PRODUCT_SERVICE_URL=http://localhost:8082
+ASSET_SERVICE_URL=http://localhost:8083
+
+# Service Configuration
+EUREKA_SERVER_INSTANCE_HOST=localhost
+EUREKA_SERVER_INSTANCE_PORT=8761
+GATEWAY_SERVICE_HOST=localhost
+GATEWAY_SERVICE_PORT=8080
