+++ .github/CODEOWNERS
@@ -1,0 +1,7 @@
+* @erdemerbaba
+/asset-service/ @lead
+/management-app/ @lead
+/user-service/ @lead
+/product-service/ @lead
+/eureka-server/ @lead
+/gateway/ @lead
