+++ .github/ISSUE_TEMPLATE/question.md
@@ -1,0 +1,11 @@
+---
+name: Question
+description: Ask a question about the project
+labels: question
+---
+
+**What is your question?**
+A clear and concise question about the project.
+
+**Additional context**
+Add any other context or screenshots related to your question here.
