+++ DEPENDENCIES.md
@@ -1,0 +1,34 @@
+# Dependencies
+
+## backend-services
+- **Frameworks**:
+  - Spring Boot
+- **Database**:
+  - MongoDB
+- **Libraries**:
+  - Lombok: Simplifies Java code with annotations.
+  - Spring Data MongoDB: Provides MongoDB integration.
+
+## management-app
+- **Frameworks**:
+  - React
+- **Libraries**:
+  - Axios: For HTTP requests.
+  - React Router: For routing.
+  - Redux: For state management.
+- **Build Tools**:
+  - Webpack: For bundling assets.
+  - Babel: For JavaScript transpilation.
+
+## eureka-server
+- **Frameworks**:
+  - Spring Cloud Netflix Eureka
+- **Libraries**:
+  - Spring Boot Actuator: For monitoring and management.
+
+## gateway
+- **Frameworks**:
+  - Spring Cloud Gateway
+- **Libraries**:
+  - Resilience4j: For fault tolerance.
+  - Spring Security: For authentication and authorization.
