+++ GOVERNANCE.md
@@ -1,0 +1,16 @@
+# Governance
+
+## Decision-Making Process
+- Decisions are made by consensus whenever possible.
+- If consensus cannot be reached, a vote will be conducted among maintainers.
+- Major changes require approval from at least 2/3 of the maintainers.
+
+## Roles and Responsibilities
+- **Maintainers**:
+  - Review and merge pull requests.
+  - Ensure the project adheres to its goals and standards.
+  - Manage releases and versioning.
+- **Contributors**:
+  - Submit pull requests for new features or bug fixes.
+  - Report issues and suggest improvements.
+  - Participate in discussions and reviews.
