+++ asset-service/src/main/java/com/management/assetservice/controller/AssetController.java
@@ -15,7 +15,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-@CrossOrigin(origins = "*", maxAge = 5800)
+@CrossOrigin(origins = "http://localhost:3000") 
 @RestController
 @RequestMapping("/api/v1/assets")
 public class AssetController {
