+++ product-service/src/main/java/com/management/productservice/controller/ProductController.java
@@ -12,7 +12,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-@CrossOrigin(origins = "*", maxAge = 5800)
+@CrossOrigin(origins = "http://localhost:3000") 
 @RestController
 @RequestMapping("/api/v1/products")
 public class ProductController {
